requirements.txt
aiohttp
async-timeout
asyncpg
attrs
beautifulsoup4
bs4
certifi
chardet
charset-normalizer
click
colorama
docopt
discord
flake8
idna
iniconfig
markovify
matplotlib
mccabe
millify
multidict
numpy
packaging
pandas
Pillow
pipreqs
pluggy
py
pycodestyle
pyflakes
pyparsing
pytest
python-dateutil
python-dotenv
pytz
requests
six
soupsieve
tomli
typing-extensions
Unidecode
urllib3
wincertstore
yarg
yarl